+++ examples/single_call/src/lib.rs
@@ -1,9 +1,7 @@
 #![cfg_attr(not(feature = "export-abi"), no_main)]
 extern crate alloc;
 
-use stylus_sdk::{
-    abi::Bytes, alloy_primitives::Address, prelude::*, stylus_core::calls::context::Call,
-};
+use stylus_sdk::{abi::Bytes, alloy_primitives::Address, prelude::*, stylus_core::calls::Call};
 
 #[storage]
 #[entrypoint]
